♻️ Refactor the response classes
